NotaryInfo refactored as operation
